refactor(pv): make a copy of props.poll with cloneDeep
mixing behaviour of props.poll and v.poll in the view causes confusion
as to which value is actually being used in the view

view sees: props.poll -> poll, v.poll -> poll
which takes precedence?  it's unclear

due to this, it's possible that code referencing props.poll OR
referencing v.poll can influence data displayed in the view

this refactor specifies that we want v.poll (renamed to v.pollCopy) used
in the view instead of props.poll itself

props.poll is used to watch for changes in the poll when user logs in;
etc. and the poll is updated